update vukan backend spv

+++ tests/vulkantest.cpp
@@ -2,8 +2,9 @@
 #include <print>
 
 int main() {
-    Tensor a = Tensor::Fill({3,100,100},3.1415f,DataType::FLOAT32);
+    Tensor a = Tensor::Fill({3,100,100},3.1415,DataType::FLOAT16);
 
     ops::println(a);
+
     return 0;
 }
